Add files via upload
I used the file SBC Shop provided and had to rewrite the code for the file Ultra.py, and renamed it. I removed the calls for Motor 3, Motor 4, all the calls for the led arrows.  I also found an undefined variable called pulse_start and defined it with a value of 0 before the while-loop for the IR sensor.  Currently,  it only moves forward and backward. I am still working on moving the head Left and Right in a consistent way. I may have to replace the head motor with a stepper motor or just add a stepper motor on top of the drone instead of using the motor in the head. 

Here is the link to the manufacture's code: https://github.com/sbcshop/MotorShield.